refactor(customers): split AdminCustomers (982→82) em hook + componentes
Extrai verbatim de src/pages/AdminCustomers.tsx (sobre a main já com o bugfix
de paginação do #309):
- useAdminCustomers: 2 infinite queries (clientes paginados + employee ids),
  loads (categorias/scores paginados/ferramentas/pedidos) e handlers
- CustomerListView (segmentos/busca/filtros/tabela densa/scroll infinito),
  Customer360View (header/KPIs/contato/scores/360/tabs), RequiresPoToggle,
  cards (MetricCard/ScoreItem), types.ts, config.ts (fmt/HEALTH_CLASSES/formatDocument)
- +8 testes (config + cards + CustomerListView)

Comportamento preservado 1:1. tsc baseline=0, eslint=0 erros
(2 warnings exhaustive-deps pré-existentes, relocados pro hook).
